+++ src/main/kotlin/caiqichang/mybatisuite/SqlLogToolWindowFactory.kt
@@ -1,4 +1,4 @@
-package caiqichang.mybatisuite.sqllog
+package caiqichang.mybatisuite
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow